Expose shared serial bundle save helper
